# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TRS/OEE DPI Pharmaceutical tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /production-entries:
    get:
      operationId: listProductionEntries
      tags: [production-entries]
      summary: List production entries with filters
      parameters:
        - name: equipmentId
          in: query
          schema:
            type: string
            format: uuid
        - name: productId
          in: query
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, validated, rejected]
        - name: shift
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionEntryWithDetails"
    post:
      operationId: createProductionEntry
      tags: [production-entries]
      summary: Create a production entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductionEntryBody"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntryWithDetails"
  /production-entries/{id}:
    get:
      operationId: getProductionEntry
      tags: [production-entries]
      summary: Get a production entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntryWithDetails"
        "404":
          description: Not found
    patch:
      operationId: updateProductionEntry
      tags: [production-entries]
      summary: Update a production entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductionEntryBody"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntryWithDetails"
  /production-entries/{id}/submit:
    post:
      operationId: submitProductionEntry
      tags: [production-entries]
      summary: Submit entry to supervisor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntryWithDetails"
  /production-entries/{id}/validate:
    post:
      operationId: validateProductionEntry
      tags: [production-entries]
      summary: Validate or reject a production entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateEntryBody"
      responses:
        "200":
          description: Entry validated/rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntryWithDetails"

  # Downtime events
...
components:
  schemas:
...
    DowntimeEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entryId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        categoryCode:
          type: ["string", "null"]
        categoryLabel:
          type: ["string", "null"]
        startTime:
          type: string
          description: HH:MM time string
        endTime:
          type: string
          description: HH:MM time string
        durationMinutes:
          type: integer
        comment:
          type: ["string", "null"]
        isDeleted:
          type: boolean
      required: [id, entryId, categoryId, startTime, endTime, durationMinutes, isDeleted]
...
    ProductionEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        equipmentId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        batchNumber:
          type: string
        shift:
          type: string
        shiftStart:
          type: string
          description: HH:MM
        shiftEnd:
          type: string
          description: HH:MM
        quantityProduced:
          type: integer
        quantityConforming:
          type: integer
        quantityRejected:
          type: integer
        status:
          type: string
          enum: [draft, submitted, validated, rejected]
        operatorId:
          type: string
          format: uuid
        supervisorId:
          type: ["string", "null"]
        supervisorComment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, date, equipmentId, productId, batchNumber, shift, shiftStart, shiftEnd, quantityProduced, quantityConforming, quantityRejected, status, operatorId, createdAt, updatedAt]
    ProductionEntryWithDetails:
      allOf:
        - $ref: "#/components/schemas/ProductionEntry"
        - type: object
          properties:
            equipmentName:
              type: ["string", "null"]
            productName:
              type: ["string", "null"]
            operatorName:
              type: ["string", "null"]
            downtimeEvents:
              type: array
              items:
                $ref: "#/components/schemas/DowntimeEvent"
            trsMetrics:
              $ref: "#/components/schemas/TrsMetrics"
    CreateProductionEntryBody:
      type: object
      properties:
        date:
          type: string
          format: date
        equipmentId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        batchNumber:
          type: string
        shift:
          type: string
        shiftStart:
          type: string
        shiftEnd:
          type: string
        quantityProduced:
          type: integer
        quantityConforming:
          type: integer
        quantityRejected:
          type: integer
      required: [date, equipmentId, productId, batchNumber, shift, shiftStart, shiftEnd, quantityProduced, quantityConforming, quantityRejected]
    UpdateProductionEntryBody:
      type: object
      properties:
        date:
          type: string
          format: date
        equipmentId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        batchNumber:
          type: string
        shift:
          type: string
        shiftStart:
          type: string
        shiftEnd:
          type: string
        quantityProduced:
          type: integer
        quantityConforming:
          type: integer
        quantityRejected:
          type: integer
    ValidateEntryBody:
      type: object
      properties:
        action:
          type: string
          enum: [validate, reject]
        comment:
          type: string
      required: [action]

    # TRS Metrics
    TrsMetrics:
      type: object
      properties:
        tT:
          type: number
          description: Temps total calendrier (minutes)
        tO:
          type: number
          description: Temps d'ouverture (minutes)
        tR:
          type: number
          description: Temps requis (minutes)
        tF:
          type: number
          description: Temps de fonctionnement (minutes)
        tN:
          type: number
          description: Temps net (minutes)
        tU:
          type: number
          description: Temps utile (minutes)
        DO:
          type: number
          description: Disponibilite operationnelle (0-1)
        TP:
          type: number
          description: Taux de performance (0-1)
        TQ:
          type: number
          description: Taux qualite (0-1)
        TRS:
          type: number
          description: TRS = DO x TP x TQ (0-1)
        TRG:
          type: number
          description: TRG = tU / tO (0-1)
        TRE:
          type: number
          description: TRE = tU / tT (0-1)
        plannedDowntimeMinutes:
          type: number
        unplannedDowntimeMinutes:
          type: number
        cadenceGap:
          type: number
          description: Gap between validated and actual cadence
      required: [tT, tO, tR, tF, tN, tU, DO, TP, TQ, TRS, TRG, TRE, plannedDowntimeMinutes, unplannedDowntimeMinutes, cadenceGap]

    # Dashboard types
...